WIP: Implement a basic CRUD API server
This includes data structs and helpers for interacting with the AddressBook's. An API server is setup with dependency injection which can be extended to include a db connection. The server has a single routes.go file where all routes are located for easier retrieval/searching. Middleware sets the Content-Type to application/json. Guard route catch all handles broken, missing or erroneous paths. To and From json utility functions are used for easier encode/decode in the routes/views.

Not done:
- put/update
- tests
- csv export